requirements.txt
numpy==1.21.5
fastapi==0.82.0
uvicorn==0.18.3
requests==2.28.2
pandas==1.3.5
openpyxl==3.1.2
python-dotenv==0.21.1
geopandas==0.10.2
pysurveycto==0.0.13
# google-cloud-documentai==2.20.2
Shapely==1.8.0
pytz==2021.3
biopython==1.81
Pillow==9.5.0